pyproject.toml
[project]
name = "multi-agent"
version = "0.1.0"
description = "Multi-Agent Slack Bot System — lightweight multi-bot chat powered by Claude"
requires-python = ">=3.11"
dependencies = [
    "slack-bolt>=1.27.0",
    "slack-sdk>=3.41.0",
    "claude-agent-sdk",
    "anthropic>=0.40.0",
    "python-dotenv>=1.0.0",
    "aiohttp>=3.13.5",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project.scripts]
multi-agent = "run:main"
